requirements/base.txt
alembic==1.10.3
anyio==3.6.2
click==8.1.3
dnspython==2.3.0
email-validator==2.0.0.post1
fastapi==0.95.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
importlib-metadata==6.4.1
importlib-resources==5.12.0
Mako==1.2.4
MarkupSafe==2.1.2
passlib==1.7.4
psycopg2-binary==2.9.6
pydantic==1.10.7
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.9
starlette==0.26.1
typing-extensions==4.5.0
uvicorn==0.21.1
zipp==3.15.0
